back/requirements.txt
aioredis==2.0.1
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Authlib==1.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
databases==0.6.1
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.1
fastapi==0.85.0
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.1
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
tomli==2.0.1
types-cryptography==3.3.23.1
typing-extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
zipp==3.10.0
